+++ Team32/Code32/src/spa/src/SP/Node/ProgramNode.cpp
@@ -23,7 +23,7 @@ bool SP::Node::ProgramNode::equals(const shared_ptr<ProgramNode>& object) {
     if (this->procedures.size() != object->procedures.size()) {
         return false;
     }
-    for (int i = 0; i < this->procedures.size(); i++) {
+    for (size_t i = 0; i < this->procedures.size(); i++) {
         if (!this->procedures[i]->equals(object->procedures[i])) {
             return false;
         }
